// src/types/theme.ts
export type Theme = {
  background: string;
  foreground: string;
  card: string;
  primary: string;
  primaryForeground: string;
  secondary: string;
  secondaryForeground: string;
  accent: string;
  accentForeground: string;
  border: string;
  radius: string;
};

export type ThemeMode = "dark" | "light";